Add all_collect to page in full member list
all_collect is faster than stream when all members have to be collected since we can run 10 fetch requests at the same time.

Also use usize for query parameters.

+++ mailchimp/src/batches.rs
@@ -118,8 +118,8 @@ pub enum BatchStatus {
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct BatchesQuery {
     pub fields: String,
-    pub count: u32,
-    pub offset: u32,
+    pub count: usize,
+    pub offset: usize,
 }
 
 #[derive(Serialize, Deserialize, Debug)]
